Use setuptools 18.2 in bootstrap.
We are installing this version with pip and pinning it in the buildout,
so we should use it while bootstrapping as well.
Otherwise when running 'bin/buildout -N' this fails with a VersionConflict:
it has 18.3.1 in bin/buildout and and the script is called in non-newest mode,
so it cannot update itself.
Alternatively, simply run bin/buildout without '-N',
seems fine.
Again, alternatively, remove the setuptools pinning. 18.3 had a problem, but 18.3.1 is fine again.