Made xpath expression in test less restrictive.
This test was looking for a section with an id, but we should just look for an id.

There is a pull request in CMFPlone for changing the section back to a div,
and the tests would fail if that gets merged.
See https://github.com/plone/Products.CMFPlone/issues/1431