Fixed unstable robot test for location criterion.
Especially the relative path criterion sometimes failed on Jenkins during this test:

  And the collection should not contain 'Document outside Folder'

Locally for me, it *always* failed.  When debugging, it appeared that
the addition of the relative path criterion did not finish before
clicking Save.

The absolute path criterion test usually passed, and this was likely
because this had a pause of one second before clicking save on the
collection edit form.  But this sometimes failed too, at least locally.

Solution is now to always do this:
- pause half a second
- change the focus (we choose the 'Sort reversed' checkbox)
- pause another half second
- then click Save

This does the trick locally, and hopefully on Jenkins too.