Fixed folder listing tests, which show more image titles.
In these tests, originally the images were already shown, but they were missing a title.
With PLIP 1483 (Retina) the 'tag' function is used in more places, which results in more titles.

For example, old:

  <img class="image-tile"
      src="http://localhost:55001/plone/robot-test-folder/test_image/@@images/image/tile" />

New:

  <img width="62" height="64" class="image-tile" title="Test Image" alt="Test Image"
      src="http://localhost:55001/plone/robot-test-folder/test_image/@@images/cbccdd2f-1a93-4935-b46f-d9e07e09cff6.png" />

Note that most images here are shown twice: in the navigation portlet and in the main content area.