Improve _replace_variable_by_path method.
At this moment we use '/'.join(obj.getPhysicalPath()) to replace
${navigation_root_url} and ${portal_url} in Link redirection but with
production site, it seems use acquisition to get correct link.

Example for same portal in production and in development:

production
https://mysite.com/my-folder/my-link
with remoteUrl = ${navigation_root_url}/my-other-folder

in local
http://localhost:8080/plone/my-folder/my-link
with same remoteUrl

generated link in local is good (http://localhost:8080/plone/my-other-folder/)
but in production, it's not correct, indeed it's
https://mysite.com/plone/my-other-folder/, and it should be
https://mysite.com/my-other-folder/

'plone' path should not be in link.
It maybe works with acquisition, but it's not correct .
So I simply replace '/'.join(obj.getPhysicalPath()) by
obj.absolute_url()

Is there good reason to use getPhysicalPath instead of absolute_url ?