Fixed robot test for collection album view.
'Test Image' used to appear three times:
- in the navigation
- as its own
- as image of the Test Folder
The first two are still there.
The last usually not, because with our leadimage changes, the Test News Item is usually picked as sample image for the Test Folder.
Which image is picked for that, does not seem entirely stable, but that is okay.
We look only for 'Test Image' as image title in a photoAlbumEntry which is not also a photoAlbumFolder.
That is what we already did for the other images, which also had a chance of being picked as sample image of their folder.