Fixed FutureWarning: Possible set difference, in tests.
This is an a regular expression.

Easy example:

```
$ python3
>>> import re
>>> re.compile(r'[a-z0-9--]+').match('a-1')
__main__:1: FutureWarning: Possible set difference at position 7
<re.Match object; span=(0, 3), match='a-1'>
```

Problem is probably that the dash is sometimes used as to indicate a-to-z or 0-to-9,
and then used as itself, where I guess '--' is meant as an escaped version of '-'.
It works, but may fail in the future.

Better is:

```
>>> re.compile(r'[a-z0-9\-]+').match('a-1')
<re.Match object; span=(0, 3), match='a-1'>
```

or actually even simpler:

```
>>> re.compile(r'[a-z0-9-]+').match('a-1')
<re.Match object; span=(0, 3), match='a-1'>
```